Organized Media related enums into Media.dart
