+++ lib/features/file_parsing/file_output.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:verticrowlabs_media_converter/features/file_parsing/file_parser.dart';
 import 'package:verticrowlabs_media_converter/features/file_parsing/file_parsing_barrel.dart';
-import 'package:verticrowlabs_media_converter/features/media_conversion/container_type.dart';
+import 'package:verticrowlabs_media_converter/features/media_conversion/media.dart';
 
 ///Takes file path and file name and returns
 ///new filepath with new name if file name is
